修改护理页面

+++ jims-service/src/main/java/com/jims/emr/enterHospital/bo/ElectronEnterHospitalBo.java
@@ -47,7 +47,7 @@ public  class ElectronEnterHospitalBo extends CrudImplService<ElectronEnterHospi
                 if (emrDiagnosisList.size() > 0) {
                     for (int i = 0; i < emrDiagnosisList.size(); i++) {
                         EmrDiagnosis diagnosis = emrDiagnosisList.get(i);
-                        if(diagnosis.getId()!=null||!"".equals(diagnosis.getId())){
+                        if(diagnosis.getId()!=null && !"".equals(diagnosis.getId())){
                             emrDiagnosisDao.delete(diagnosis.getId());
                             diagnosis.setId("");
                         }
@@ -55,7 +55,6 @@ public  class ElectronEnterHospitalBo extends CrudImplService<ElectronEnterHospi
                         diagnosis.setParentId("0");
                         diagnosis.setDiagnosisDoc(loginInfo.getPersionId());
                         diagnosis.setOrgId(loginInfo.getOrgId());
-
                         diagnosis.setItemNo(i+1);
                         try {
                             if (diagnosis.getIsNewRecord()) {
@@ -66,6 +65,7 @@ public  class ElectronEnterHospitalBo extends CrudImplService<ElectronEnterHospi
                                 num = emrDiagnosisDao.update(diagnosis);
                             }
                         } catch (Exception e) {
+                            e.printStackTrace();
                             return num + "";
                         }
                         return num + "";
